EVRAZBACKEND/requirements.txt
django==4.1.7
kafka-python==2.0.2
pytz==2022.7.1